:star: Accomplished witch bouncing

+++ LeapOfFaith/Block.h
@@ -6,11 +6,11 @@
 class Block:public Cube
 {
 public:
-	int width;
-	int  height;
+	float width;
+	float  height;
 	int num;
 public:
-	Block(CBitmap *,int x,int y,int num,int w,int h);
+	Block(CBitmap *, float x, float y, int num, float w, float h);
 	~Block();
 
 private:
@@ -18,7 +18,7 @@ private:
 };
 
 
-inline Block::Block(CBitmap *bmp, int x, int y, int num,int w = BLOCK_WIDTH, int h=BLOCK_HEIGHT):Cube(bmp)
+inline Block::Block(CBitmap *bmp, float x, float y, int num, float w = BLOCK_WIDTH, float h=BLOCK_HEIGHT):Cube(bmp)
 {
 	this->x = x;
 	this->y = y;
